Use longs instead of IntPtrs

+++ src/Memory/DmaMemory.cs
@@ -4,10 +4,10 @@ namespace IxyCs.Memory
 {
     public class DmaMemory
     {
-        public readonly IntPtr VirtualAddress;
+        public readonly long VirtualAddress;
         public readonly long PhysicalAddress;
 
-        public DmaMemory(IntPtr virt, long phys)
+        public DmaMemory(long virt, long phys)
         {
             VirtualAddress = virt;
             PhysicalAddress = phys;
